+++ webjam_qt/__init__.py
@@ -11,4 +11,4 @@ then will be removed.
 
 __all__ = ["__version__"]
 
-__version__ = "0.27.1"
+__version__ = "0.27.2"